+++ response.go
@@ -147,8 +147,10 @@ func (resp *Response) decodeBody() (err error) {
 		offset := resp.buf.Offset()
 		defer resp.buf.Seek(offset)
 
-		var l int
+		var l, larr int
 		var stmtID, bindCount uint64
+		var serverProtocolInfo ProtocolInfo
+		var feature ProtocolFeature
 
 		d := newDecoder(&resp.buf)
 
@@ -190,6 +192,22 @@ func (resp *Response) decodeBody() (err error) {
 				if bindCount, err = d.DecodeUint64(); err != nil {
 					return err
 				}
+			case KeyVersion:
+				if err = d.Decode(&serverProtocolInfo.Version); err != nil {
+					return err
+				}
+			case KeyFeatures:
+				if larr, err = d.DecodeArrayLen(); err != nil {
+					return err
+				}
+
+				serverProtocolInfo.Features = make([]ProtocolFeature, larr)
+				for i := 0; i < larr; i++ {
+					if err = d.Decode(&feature); err != nil {
+						return err
+					}
+					serverProtocolInfo.Features[i] = feature
+				}
 			default:
 				if err = d.Skip(); err != nil {
 					return err
@@ -204,6 +222,18 @@ func (resp *Response) decodeBody() (err error) {
 			}
 			resp.Data = []interface{}{stmt}
 		}
+
+		// Tarantool may send only version >= 1
+		if (serverProtocolInfo.Version != ProtocolVersion(0)) || (serverProtocolInfo.Features != nil) {
+			if serverProtocolInfo.Version == ProtocolVersion(0) {
+				return fmt.Errorf("No protocol version provided in Id response")
+			}
+			if serverProtocolInfo.Features == nil {
+				return fmt.Errorf("No features provided in Id response")
+			}
+			resp.Data = []interface{}{serverProtocolInfo}
+		}
+
 		if resp.Code != OkCode && resp.Code != PushCode {
 			resp.Code &^= ErrorCodeBit
 			err = Error{resp.Code, resp.Error}